agent/requirements.txt
ollama==0.4.7
requests==2.32.3
python-docx==1.1.2
